requirements.txt
Django>=4.0
djangorestframework
easyocr
torch
torchvision
psycopg2-binary
